fix another place that can lead to differences on 32vs64 bits

+++ src/mip/HighsCutGeneration.cpp
@@ -56,28 +56,29 @@ bool HighsCutGeneration::determineCover(bool lpSol) {
 
   // sort the remaining variables by the contribution to the rows activity in
   // the current solution
-  std::sort(cover.begin() + coversize, cover.begin() + maxCoverSize,
-            [&](HighsInt i, HighsInt j) {
-              double contributionA = solval[i] * vals[i];
-              double contributionB = solval[j] * vals[j];
+  std::sort(
+      cover.begin() + coversize, cover.begin() + maxCoverSize,
+      [&](HighsInt i, HighsInt j) {
+        double contributionA = solval[i] * vals[i];
+        double contributionB = solval[j] * vals[j];
 
-              // for equal contributions take the larger coefficients first
-              // because this makes some of the lifting functions more likely to
-              // generate a facet
-              if (std::abs(contributionA - contributionB) <= feastol) {
-                // if the value is equal too, choose a random tiebreaker based
-                // on hashing the column index and the current number of pool
-                // cuts
-                if (std::abs(vals[i] - vals[j]) <= feastol)
-                  return HighsHashHelpers::hash(
-                             std::make_pair(inds[i], cutpool.getNumCuts())) >
-                         HighsHashHelpers::hash(
-                             std::make_pair(inds[j], cutpool.getNumCuts()));
-                return vals[i] > vals[j];
-              }
+        // for equal contributions take the larger coefficients first
+        // because this makes some of the lifting functions more likely to
+        // generate a facet
+        if (std::abs(contributionA - contributionB) <= feastol) {
+          // if the value is equal too, choose a random tiebreaker based
+          // on hashing the column index and the current number of pool
+          // cuts
+          if (std::abs(vals[i] - vals[j]) <= feastol)
+            return HighsHashHelpers::hash(std::make_pair(
+                       uint32_t(inds[i]), uint32_t(cutpool.getNumCuts()))) >
+                   HighsHashHelpers::hash(std::make_pair(
+                       uint32_t(inds[j]), uint32_t(cutpool.getNumCuts())));
+          return vals[i] > vals[j];
+        }
 
-              return contributionA > contributionB;
-            });
+        return contributionA > contributionB;
+      });
 
   const double minlambda =
       std::max(10 * feastol, feastol * std::abs(double(rhs)));
